C++做Server，C#做Client测试成功

+++ EmbedXrpc.CSharp.NetCore/EmbedXrpcCommon.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace EmbedXrpc
 {
-    public enum ResponseState:UInt32
+    public enum ResponseState:Byte
     {
         Ok=1,
         Timeout=2,
